move most of integtest into a callable workflow
Rather than integtest.yaml containing all of the integ test logic, it
now just invokes a new workflow, which contains all of its old logic.
This makes it possible to invoke the tests from other repos.

It seems silly to create a workflow that does nothing but call another
workflow, but I _think_ that's necessary because during a cron-triggered
run, the `inputs.*` object is empty (not populated with the default
values).

It's also silly and a bit brittle to have the workflow_dispatch and
workflow_call sections duplicate the same inputs, but unfortunately that
seems to be what GH Actions require.

As part of this, rename integtest.yaml to nightly-integ-tests.yaml. This
will kill the existing workflow history, but there's not much of it
right now anyway.

Also, ignore ts-eks-helm for now (see #49).