refactor main.go
This just wraps all of the stuff into a lightweight struct that lets you
configure various aspects of main. This lets us create other "editions"
of klotho (and specifically pro) without having to maintain a separate
copy of this main.go, which can introduce drift.

The `main.go` changes contain a bunch of - (and corresponding + in
`klothomain.go`), but it's pretty much a lift-and-shift. A few places
had some `foo` turning into a `km.Foo` or similar, but that's about it.

After that, just a bit of light refactoring/moving to untangle a couple
cyclic dependencies in `analytics` and `cli`.