remove client-side login retry
I'm also moving the Login error handler to a callback, just as an
organizational tool. We want to send the analytics for any error
manually (since the client hasn't been set up yet), but the Login method
shouldn't have to know about that intricacy. klothomain.go does, so
that's where we should put that logic