remove client-side login retry
Before this, the server's /logintoken API would return instantly, with
either the token from auth0 or 404 if it hadn't been set yet from the
auth0.com callback. So, we needed a callback not just for intermittent
errors, but for the happy path functionality. Now, the /logintoken call
hangs until it gets that token (or times out), so the happy path doesn't
need a retry. We could still need a retry for intermittent failurs, but
that's a more general problem that we should resolve across all http
calls, not just for auth.

I'm also moving the Login error handler to a callback, just as an
organizational tool. We want to send the analytics for any error
manually (since the client hasn't been set up yet), but the Login method
shouldn't have to know about that intricacy. klothomain.go does, so
that's where we should put that logic