fail-open on auth failures
If the `credentials.json` file isn't there at all, then require a login.
But if it's there but the login doesn't work for whatever reason, just
issue a warning and move on. Meanwhile, in the login path, handle errors
by writing an empty `credentials.json` and then ignoring the error.

Relatedly, add a workaround for #195. It's very easy to hit it with this
new path, so I felt like a quick workaround is in order.