fix pin-and-downgrade
Previously, if you tried to pin to a lower version, we would treat it as
a no-op (and keep you at your current version). Now, we'll perform the
update ("downdate").

This resolves #187.

We've had a few back-and-forths in this code, most recently #137. I
added unit tests to try and put this to bed for good, or at least
prevent further regressions.

I found myself getting kinda confused about the various situations, so I
wrote a novel to explain it to myself. I figured I'd keep it around for
the next person.