send properties separately from event name in logs
When we hook into the log entries, don't send the sanitized fields as
json in the event name. Instead, just send the log level for the event
name (e.g. "WARN"), and send each sanitized field as a property with a
dot-delimited key. For example, a file field would send:

    log.file.extension: .json
    log.file.path: sha256:<snip>

This resolves #239.