improve auth analytics
Always get a LoginInfo from the authorizer; this is never nil, and
always contains an "Authorizer" field which should be a non-empty string
describing the authorizer (for now, just "auth0" or "local"). Then,
always attach that info to the analytics. This will let us know e.g. how
many compiles are using local vs auth0.

This resolves #218