allow cycles in the graph
This uses dominikbraun/graph for the underlying graph. I'm not sure it's
really what we want, because it doesn't let us easily graph the graph's
roots, or the adjecent edges/vertices for a given vertex. But it's good
enough for now, and we can always put in a new backing graph later.