basic graph structs (#383)
This uses dominikbraun/graph for all the real calculations, and just
introduces a thin wrapper around that. The wrapper provides a simpler
facade for the DAG and lets us swap out for other DAG implementations,
if we want to do that later.


I'm not sure dominikbraun/graph is really what we want, because it
doesn't let us easily graph the graph's roots, or the adjecent
edges/vertices for a given vertex. But it's good enough for now, and we
can always put in a new backing graph later. We'll know more about our
graph's requirements as we get more into the project.