v1 of the iac plugin (#394)
The "main event" here is templatesCompiler. It takes in a graph
(expected to be a DAG) of cloud resources, and translates each one into
a `const varname = ....` statement. For now, the idea is they'd all go
into a single `index.ts` -- but we can think about splitting those up
later.