write cpu and memory limits to k8s manifest (#479)
Both for deployments and for pods. This includes a bit of refactoring
work, to combine the "make sure there's no more than 1 container, and
then configure it" that both deploy and pod were doing. That approach
lets us test that configuration separately from the spec that's being
configured, which means we don't need to do the full combination of all
possible inputs for both deploy and pod.

In service of #419.