add cert-manager and metrics-server to k8s (#507)
As part of this, I'm also changing the HelmChart.Values map to have
values of type `any`, instead of `IaCValue`. This because the
cert-manager chart wants values that include a boolean (for
`installCRDs`) and a nested Record/map (for `webhook`). In order to
accomplish that, I also changed the helm_chart/factory.ts Args to take
`pulumi.Output<any>` instead of `...<string>`. This is what the pulumi
field is actually typed as, so we don't actually lose any checks by
doing it this way.

Also fix a typo in the factory.ts's handling of args.Repo.

Also also provider a prettier error in coretesting, because otherwise
it's very hard to read the diff when two lists have different elements.

in service of #420