rm struct field "manifestKind"
We can reflectively get it from the K, so we don't actually need it as a
struct field.

We could do the same with `fieldToTransform` in theory, assuming the
field name is always the K's type name (which is the case today) -- but
that means that we could only catch any mismatch/errors at runtime.