use new visualizer for generating the graph
Mostly what it says on the tin; a couple helper structs and small
refactoring along the way.

One thing to note is TestAllTypesHaveIcons. It gets skipped unless you
set an env var KLOTHO_VIZ_BASE_URL. Once klothoplatform/visualizer#10
gets merged, we may want to set this env var in our GH Action so that
this test runs, hitting the public server.

In service of #499