add S3Bucket.Create, and factor out a test util
The bucket create is pretty simple, because it doesn't have any
dependencies.

I realized that its test was almost identical to that of
LamdbaFunction.Create, so I factored out a helper that I hope will
provide a nice pattern for testing these as we write them. I backported
the LambdaFunction's test as a check that there really is commonality,
but kept the rest unchanged just in the interest of time (and smaller
diff).