checkpoint (NOT WORKING)
This tries to fix #653 by deferring the stage name. Unfortunately, the
deployment doesn't have a stage at `pulumi up` time, so the origin path
becomes "/UNSPECIFIED_STAGE" (literally that text).

I think to fix this, we would need to either:

- set the [Deployment's stageName][1]; or
- explicitly [create a State][2]

In either case, we would no longer be relying on the implicit stage
creation-and-naming.

[1]: https://www.pulumi.com/registry/packages/aws/api-docs/apigateway/deployment/#stagename_nodejs
[2]: https://www.pulumi.com/registry/packages/aws/api-docs/apigateway/stage/